Add Layouts in Samples app
